roachtest: add placeholder type hint
Fixes: #91728, #91725, #91726, #91727, #91736

Release note: None

+++ pkg/cmd/roachtest/tests/smoketest_secure.go
@@ -66,7 +66,7 @@ func multitenantSmokeTest(ctx context.Context, t test.Test, c cluster.Cluster) {
 	tenID := 11
 	ten := createTenantNode(ctx, t, c, c.Node(1), tenID, 2, 8011, 9011)
 	runner := sqlutils.MakeSQLRunner(c.Conn(ctx, t.L(), 1))
-	runner.Exec(t, `SELECT crdb_internal.create_tenant($1)`, tenID)
+	runner.Exec(t, `SELECT crdb_internal.create_tenant($1::INT)`, tenID)
 	ten.start(ctx, t, c, "./cockroach")
 
 	// this doesn't work yet, roachprod knows nothing about tenants
